Update accounts & targets

+++ assets/js/modules/targets.js
@@ -160,23 +160,43 @@ function overviewSubjects(category){
                             class : 'btn-nobg btn-block text-muted left addTarget'
                           }),
                           targetBtnsContainer = $('<div></div>').attr('class','targetBtnsContainer'),
-                          subjectDeleteBtn = $('<button></button>')
-                            .attr('class','btn btn-nobg float-right red')
-                            .attr('style','margin-top:-38px;')
-                            .html('<i class="fas fa-times"></i>')
-                            .on('click',(event)=>{
+                          subjectDeleteBtn = component.btn({
+                              html : '<i class="fas fa-times"></i>',
+                              class : 'nobg float-right red',
+                              style : 'margin-top:-38px;', 
+                              event : ['click',() => {
                                 const subjectId = event.target.parentElement.id ? event.target.parentElement.id : event.target.parentElement.parentElement.id;
                                 component.api({
-                                  method : 'delete',
-                                  url : `api/subjects/${subjectId}`,
-                                  callback : ()=>{
-                                    overviewSubjects(category)
-                                    component.alert({
-                                        message : '<i class="fas fa-times"></i> Onderwerp verwijderd'
+                                    method : 'delete',
+                                    url : `api/subjects/${subjectId}`,
+                                    callback : ()=>{
+                                        overviewSubjects(category)
+                                        component.alert({
+                                            message : '<i class="fas fa-times"></i> Onderwerp verwijderd'
+                                        })
+                                    }
+                                })
+                              }]
+                              
+                              /*
+                              confirm : {
+                                  msg : 'Weet je zeker dat je dit onderwerp wilt verwijderen?',
+                                  placement : 'left',
+                                  confirm : () => {
+                                    const subjectId = event.target.parentElement.id ? event.target.parentElement.id : event.target.parentElement.parentElement.id;
+                                    component.api({
+                                        method : 'delete',
+                                        url : `api/subjects/${subjectId}`,
+                                        callback : ()=>{
+                                            overviewSubjects(category)
+                                            component.alert({
+                                                message : '<i class="fas fa-times"></i> Onderwerp verwijderd'
+                                            })
+                                        }
                                     })
                                   }
-                                })
-                            }),
+                              }*/
+                          }),
                           subjectOptions = $('<div></div>')
                             .attr('class','subjectOptions float-right')
                             .append(subjectDeleteBtn),
@@ -232,7 +252,8 @@ function overviewTargetLevels(target){
             $('#overviewTargetSubjectName').html(target.name);
             for(const levelIndex in levels){
                 const levelBtn = component.btn({
-                    txt : `Level ${levelIndex} : ${levels[levelIndex].name}`
+                    txt : `Level ${levelIndex} : ${levels[levelIndex].name}`,
+                    class : 'primary green'
                 })
                 $('#overviewTargetLevelsBtns').append(levelBtn)
             }
